src/requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.9.3
aniso8601==9.0.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Authlib==1.2.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Deprecated==1.2.13
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.9
frozenlist==1.3.3
gevent==22.10.2
googleapis-common-protos==1.56.2
greenlet==2.0.2
grpcio==1.51.3
hiredis==2.2.2
identify==2.5.17
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
multidict==6.0.4
nodeenv==1.7.0
opentelemetry-api==1.16.0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
opentelemetry-exporter-jaeger-thrift==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
orjson==3.8.6
packaging==23.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
rauth==0.7.3
redis==4.4.2
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.3
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.19.0
user-agents==2.2.0
Werkzeug==2.2.2
wrapt==1.15.0
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
